refactor(migrations): consolidate professional travel migrations into single file
