--- backend/alembic/versions/2026_01_09_1400-ce20d9691b22_create_locations_table.py
@@ -1,105 +1,0 @@
-"""create locations table
-
-Revision ID: ce20d9691b22
-Revises: 3d337ca43cfa
-Create Date: 2026-01-09 14:00:00.000000
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-import sqlmodel
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "ce20d9691b22"  # noqa: F841
-down_revision: Union[str, Sequence[str], None] = "3d337ca43cfa"  # noqa: F841
-branch_labels: Union[str, Sequence[str], None] = None  # noqa: F841
-depends_on: Union[str, Sequence[str], None] = None  # noqa: F841
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "locations",
-        sa.Column(
-            "created_at",
-            sa.DateTime(),
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(),
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-            nullable=False,
-        ),
-        sa.Column("created_by", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column("updated_by", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column(
-            "transport_mode",
-            sqlmodel.sql.sqltypes.AutoString(length=50),
-            nullable=False,
-        ),
-        sa.Column(
-            "name",
-            sqlmodel.sql.sqltypes.AutoString(length=255),
-            nullable=False,
-        ),
-        sa.Column("latitude", sa.Float(), nullable=False),
-        sa.Column("longitude", sa.Float(), nullable=False),
-        sa.Column(
-            "iata_code",
-            sqlmodel.sql.sqltypes.AutoString(length=10),
-            nullable=True,
-        ),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    # Create indexes
-    op.create_index(
-        op.f("ix_locations_created_by"),
-        "locations",
-        ["created_by"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_locations_iata_code"),
-        "locations",
-        ["iata_code"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_locations_name"),
-        "locations",
-        ["name"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_locations_transport_mode"),
-        "locations",
-        ["transport_mode"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_locations_updated_by"),
-        "locations",
-        ["updated_by"],
-        unique=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_locations_updated_by"), table_name="locations")
-    op.drop_index(op.f("ix_locations_transport_mode"), table_name="locations")
-    op.drop_index(op.f("ix_locations_name"), table_name="locations")
-    op.drop_index(op.f("ix_locations_iata_code"), table_name="locations")
-    op.drop_index(op.f("ix_locations_created_by"), table_name="locations")
-    op.drop_table("locations")
-    # ### end Alembic commands ###
