--- backend/alembic/versions/2026_01_15_1200-3d337ca43cfa_create_professional_travels_table.py
@@ -1,144 +1,0 @@
-"""create professional_travels table
-
-Revision ID: 3d337ca43cfa
-Revises: 99bb31e235b5
-Create Date: 2026-01-15 12:00:00.000000
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-import sqlmodel
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "3d337ca43cfa"  # noqa: F841
-down_revision: Union[str, Sequence[str], None] = "99bb31e235b5"  # noqa: F841
-branch_labels: Union[str, Sequence[str], None] = None  # noqa: F841
-depends_on: Union[str, Sequence[str], None] = None  # noqa: F841
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "professional_travels",
-        sa.Column(
-            "created_at",
-            sa.DateTime(),
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(),
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-            nullable=False,
-        ),
-        sa.Column("created_by", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column("updated_by", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column(
-            "traveler_id", sqlmodel.sql.sqltypes.AutoString(length=50), nullable=False
-        ),
-        sa.Column(
-            "traveler_name",
-            sqlmodel.sql.sqltypes.AutoString(length=255),
-            nullable=False,
-        ),
-        sa.Column(
-            "origin", sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False
-        ),
-        sa.Column(
-            "destination",
-            sqlmodel.sql.sqltypes.AutoString(length=255),
-            nullable=False,
-        ),
-        sa.Column("departure_date", sa.Date(), nullable=True),
-        sa.Column("return_date", sa.Date(), nullable=True),
-        sa.Column("is_round_trip", sa.Boolean(), nullable=False, server_default="0"),
-        sa.Column(
-            "transport_mode",
-            sqlmodel.sql.sqltypes.AutoString(length=50),
-            nullable=False,
-        ),
-        sa.Column("class", sqlmodel.sql.sqltypes.AutoString(length=50), nullable=True),
-        sa.Column("distance_km", sa.Float(), nullable=True),
-        sa.Column("number_of_trips", sa.Integer(), nullable=False, server_default="1"),
-        sa.Column("kg_co2eq", sa.Float(), nullable=True),
-        sa.Column(
-            "unit_id", sqlmodel.sql.sqltypes.AutoString(length=50), nullable=False
-        ),
-        sa.Column(
-            "provider", sqlmodel.sql.sqltypes.AutoString(length=50), nullable=True
-        ),
-        sa.Column("year", sa.Integer(), nullable=False),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    # Create indexes
-    op.create_index(
-        op.f("ix_professional_travels_created_by"),
-        "professional_travels",
-        ["created_by"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_professional_travels_unit_id"),
-        "professional_travels",
-        ["unit_id"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_professional_travels_updated_by"),
-        "professional_travels",
-        ["updated_by"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_professional_travels_year"),
-        "professional_travels",
-        ["year"],
-        unique=False,
-    )
-    # Composite index for primary filtering
-    op.create_index(
-        "ix_professional_travels_unit_id_year",
-        "professional_travels",
-        ["unit_id", "year"],
-        unique=False,
-    )
-    # Index for traveler filtering
-    op.create_index(
-        "ix_professional_travels_traveler_id",
-        "professional_travels",
-        ["traveler_id"],
-        unique=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(
-        "ix_professional_travels_traveler_id", table_name="professional_travels"
-    )
-    op.drop_index(
-        "ix_professional_travels_unit_id_year", table_name="professional_travels"
-    )
-    op.drop_index(
-        op.f("ix_professional_travels_year"), table_name="professional_travels"
-    )
-    op.drop_index(
-        op.f("ix_professional_travels_updated_by"), table_name="professional_travels"
-    )
-    op.drop_index(
-        op.f("ix_professional_travels_unit_id"), table_name="professional_travels"
-    )
-    op.drop_index(
-        op.f("ix_professional_travels_created_by"), table_name="professional_travels"
-    )
-    op.drop_table("professional_travels")
-    # ### end Alembic commands ###
